join connection and robot view
